Wrap player variables
