+++ engine/src/main/kotlin/world/gregs/voidps/engine/data/serializer/PlayerBuilder.kt
@@ -1,5 +1,6 @@
 package world.gregs.voidps.engine.data.serializer
 
+import world.gregs.voidps.engine.client.variable.Variables
 import world.gregs.voidps.engine.entity.Values
 import world.gregs.voidps.engine.entity.character.contain.Container
 import world.gregs.voidps.engine.entity.character.player.Player
@@ -23,7 +24,7 @@ internal data class PlayerBuilder(
         tile = Tile(tile),
         containers = containers,
         experience = experience,
-        variables = variables,
+        variables = Variables(variables),
         levels = Levels(levelOffsets),
         name = name,
         values = Values(values),
