+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/player/Player.kt
@@ -14,6 +14,7 @@ import world.gregs.voidps.engine.action.Suspension
 import world.gregs.voidps.engine.client.ui.InterfaceOptions
 import world.gregs.voidps.engine.client.ui.Interfaces
 import world.gregs.voidps.engine.client.ui.dialogue.Dialogues
+import world.gregs.voidps.engine.client.variable.Variables
 import world.gregs.voidps.engine.data.StorageStrategy
 import world.gregs.voidps.engine.data.serializer.PlayerBuilder
 import world.gregs.voidps.engine.delay
@@ -58,9 +59,8 @@ class Player(
     @JsonIgnore
     override val movement: Movement = Movement(),
     val containers: MutableMap<String, Container> = mutableMapOf(),
-    val variables: MutableMap<String, Any> = mutableMapOf(),
-    @JsonIgnore
-    val temporaryVariables: MutableMap<String, Any> = mutableMapOf(),
+    @get:JsonUnwrapped
+    val variables: Variables = Variables(),
     override val values: Values = Values(),
     @JsonIgnore
     val dialogues: Dialogues = Dialogues(),
